fix(InfoIcon.tsx): use FaRegQuestionCircle instead of FaQuestionCircle for better visual representation
This change updates the icon used for the InfoIcon component. The previous icon, `FaQuestionCircle`, had a filled circle, while `FaRegQuestionCircle` provides an outlined circle, which is a more common and visually appealing representation for an information icon.

Also, adds styling for the info-icon class in `tailwind-entry.css` to ensure consistent sizing and alignment of the icon.